Fix the third batch of MSVC conversion warnings
More narrowing casts in the rx layer and RandomX core, including the
blob size argument in RxVeil, plus one more anonymous catch. These kept
surfacing ten at a time because github caps annotations per job.

+++ src/crypto/randomx/superscalar.cpp
@@ -151,7 +151,7 @@ namespace randomx {
 			return name_;
 		}
 		int getSize() const {
-			return ops_.size();
+			return static_cast<int>(ops_.size());
 		}
 		bool isSimple() const {
 			return getSize() == 1;
